[GOHW-5]: Тесты

See merge request frt550/tasks!5

+++ internal/pkg/bot/command/delete/delete.go
@@ -34,7 +34,7 @@ func (c *command) Process(args string) string {
 		return "Please, enter a valid task id"
 	}
 
-	if task, err := c.task.Delete(ctx, uint(id)); err != nil {
+	if task, err := c.task.Delete(ctx, uint64(id)); err != nil {
 		return errPkg.Error(err)
 	} else {
 		return fmt.Sprintf("Task %v is deleted", task)
